created only function using fetch
