+++ src/schemas/uruguay/uruguay.ts
@@ -2,10 +2,10 @@ import { Schema, model} from 'mongoose';
 import { ICountry } from '../brazil/brazil';
 
 
-const uruguaiSchema = new Schema<ICountry>({
+const uruguaySchema = new Schema<ICountry>({
     university: { type: String, required: true}
 },
-    {collection: 'uruguai'}
+    {collection: 'uruguay'}
 );
 
-export const Uruguai = model<ICountry>('Uruguai', uruguaiSchema);
+export const Uruguai = model<ICountry>('Uruguay', uruguaySchema);
