+++ src/server.ts
@@ -1,21 +1,13 @@
 import express from 'express';
-import { Countries } from './models/brazil';
+import { Countries } from './models/fetchAll';
 import authenticate  from './database/database';
 
 const app = express();
 const database = authenticate()
 
+
 const country = new Countries;
-app.use('/', country.brazil)
-// const country = new Countries();
-// country.brazil();
-// country.argentina();
-// country.chile();
-// country.colombia();
-// country.paraguai();
-// country.peru();
-// country.suriname();
-// // country.uruguay();
+app.use('/', [country.argentina, country.brazil, country.chile, country.suriname]);
 
 try {
     app.listen(3000, () => {
